02.23

+++ javaProject/src/operators/CheckOverflowExample.java
@@ -3,7 +3,7 @@ package operators;
 public class CheckOverflowExample {
 	public static void main(String[] args) {
 		try {
-			int result = safeAdd(2000000000, 2000000000);
+			long result = safeAdd(2000000000, 2000000000);
 			System.out.println(result);
 		} catch (Exception e) {
 			System.out.println("오버플로우가 발생하여 정확하게 계산할 수 없음. " + e.getMessage());
@@ -11,12 +11,12 @@ public class CheckOverflowExample {
 		System.out.println("프로그램 종료");
 	}
 
-	private static int safeAdd(int left, int right) {
+	private static long safeAdd(long left, long right) {
 		if ((right > 0)) {
-			if (left > (Integer.MAX_VALUE - right)) {
+			if (left > (Long.MAX_VALUE - right)) {
 				throw new ArithmeticException("오버플로우 발생");
 			} else {
-				if (left < (Integer.MIN_VALUE - right)) {
+				if (left < (Long.MIN_VALUE - right)) {
 					throw new ArithmeticException("오버플로우 발생");
 				}
 			}
